backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "aiosqlite>=0.21.0",
    "bcrypt==4.0.1",
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.17",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.28",
    "langgraph>=0.5.4",
    "matplotlib>=3.10.3",
    "pandas>=2.3.1",
    "passlib>=1.7.4",
    "plotly>=6.2.0",
    "pydantic-settings>=2.10.1",
    "pydantic[email]>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "seaborn>=0.13.2",
    "uvicorn>=0.35.0",
]
